+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "ftlXml" extension will be documented in this file.
 ## 0.4.4
 - shipBlueprint [unlock] id text supported
 - add crewCount [class] auto complete
+- add weaponList [load] auto complete
 
 ## 0.4.3
 - Auto parse files when added, including png and audio files